# Use rebar3 for compilation and AVM packbeam generation.