--- a/libs/eavmlib/rebar.lock.license
+++ b/libs/eavmlib/rebar.lock.license
@@ -1,2 +1,0 @@
-SPDX-License-Identifier: CC0-1.0
-SPDX-FileCopyrightText: AtomVM Contributors