--- a/libs/esp32boot/rebar.config.in
+++ b/libs/esp32boot/rebar.config.in
@@ -1,0 +1,34 @@
+%
+% This file is part of AtomVM.
+%
+% Copyright 2023 Fred Dushin <fred@dushin.net>
+%
+% Licensed under the Apache License, Version 2.0 (the "License");
+% you may not use this file except in compliance with the License.
+% You may obtain a copy of the License at
+%
+%    http://www.apache.org/licenses/LICENSE-2.0
+%
+% Unless required by applicable law or agreed to in writing, software
+% distributed under the License is distributed on an "AS IS" BASIS,
+% WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+% See the License for the specific language governing permissions and
+% limitations under the License.
+%
+% SPDX-License-Identifier: Apache-2.0 OR LGPL-2.1-or-later
+%
+
+{erl_opts, [debug_info]}.
+{deps, []}.
+
+%% {src_dirs, ["@CMAKE_CURRENT_SOURCE_DIR@/src"]}.
+%% {root_dir, "@CMAKE_CURRENT_SOURCE_DIR@"}.
+
+{base_dir, "@CMAKE_CURRENT_BINARY_DIR@/_build"}.
+
+{plugins, [
+    atomvm_rebar3_plugin
+]}.
+{atomvm_rebar3_plugin, [
+    {packbeam, []}
+]}.