--- a/libs/estdlib/.gitignore
+++ b/libs/estdlib/.gitignore
@@ -1,0 +1,6 @@
+# Copyright 2023 Fred Dushin <fred@dushin.net>
+#
+# SPDX-License-Identifier: Apache-2.0 OR LGPL-2.1-or-later
+rebar.config
+rebar.lock
+_checkouts/*