--- a/src/platforms/rp2040/tests/test_erl_sources/.gitignore
+++ b/src/platforms/rp2040/tests/test_erl_sources/.gitignore
@@ -1,0 +1,7 @@
+# Copyright 2023 Fred Dushin <fred@dushin.net>
+#
+# SPDX-License-Identifier: Apache-2.0 OR LGPL-2.1-or-later
+rebar.config
+rebar.lock
+_checkouts/*
+src/test_crypto.erl