--- a/src/platforms/rp2040/tests/test_erl_sources/src/rp2040_test_modules.app.src
+++ b/src/platforms/rp2040/tests/test_erl_sources/src/rp2040_test_modules.app.src
@@ -1,0 +1,34 @@
+%
+% This file is part of AtomVM.
+%
+% Copyright 2020 Davide Bettio <davide@uninstall.it>
+%
+% Licensed under the Apache License, Version 2.0 (the "License");
+% you may not use this file except in compliance with the License.
+% You may obtain a copy of the License at
+%
+%    http://www.apache.org/licenses/LICENSE-2.0
+%
+% Unless required by applicable law or agreed to in writing, software
+% distributed under the License is distributed on an "AS IS" BASIS,
+% WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+% See the License for the specific language governing permissions and
+% limitations under the License.
+%
+% SPDX-License-Identifier: Apache-2.0 OR LGPL-2.1-or-later
+%
+
+{application, rp2040_test_modules, [
+    {description, "Erlang AtomVM library"},
+    {vsn, "0.1.0"},
+    {registered, []},
+    {applications, [
+        kernel,
+        stdlib
+    ]},
+    {env,[]},
+    {modules, []},
+    {maintainers, []},
+    {licenses, ["Apache 2.0"]},
+    {links, [{"GitHub", "https://github.com/atomvm/AtomVM/"}]}
+ ]}.