--- a/tests/erlang_tests/tools/to_hrl.erl
+++ b/tests/erlang_tests/tools/to_hrl.erl
@@ -1,6 +1,3 @@
-#!/usr/bin/env escript
-%% -*- erlang -*-
-%%! -sname to_hrl debug verbose
 %
 % This file is part of AtomVM.
 %