This PR changes the build of Erlang libraries and tests to use rebar3 as a build tool, instead of calling the Erlang compiler directly.  The `atomvm_rebar3_packbeam` rebar3 plugin is used to create PackBEAM (`.avm`) files, instead of the C-based `PackBEAM` tool.

This PR includes the following changes:

* The erlang libraries in the `libs` directory have been modifed to comply with `rebar3` build conventions, including a src directory, an `<application-name>.app.src` file for each library, and a `rebar.config.in` file, for which CMake rules will generate a `rebar.config` with proper paths.  Many of the changes in this directory also relate simply to file moves, in order to comply with `rebar3` build conventions.
* The `alisp` project has been folded into the `eavmlib` project, as we have no use case for building or deploying these modules separately.
* The erlang tests, including unit tests and the `estdlib` and `eavmlib`, have been modified to also use `rebar3`.  Many of the changes in this directory relate simply to file moves, in order to comply with `rebar3` build conventions.
* The Elixir builds have not been significantly modified, for example to use the mix tool and `ExAtomVM` plugin.  That work can come later, once the `ExAtomVM` tool reaches functional parity with the `atomvm_rebar3_plugin`.
* The Erlang and Elixir examples have not been changed to use the new build system, as we plan to deprecate these examples in favor of the `atomvm_examples` repository in the future.
* The rp2040 test build has been modified to use the `rebar3` build system, though some of the logic has been copied from the base build system.  Due to a potentially missing feature in the `atomvm_rebar3_plugin` `uf2create` task, we have left use of the `uf2tool.erl` escript.  A TODO has been left to identify the issue.
* Many of the Github workflows have been modified to include and build `rebar3` as well as various adjustments to resulting paths.
* Several TODOs have been identified in the existing and new build system and have been identified with comments.

## Benchmarks

The following benchmarks were run on MacOS, FreeBSD, and Linux systems (though the Linux system is running virtualized on FreeBSD, so numbers there should be taken with a grain of salt).  Comparisons are made between build times on the `master` and this branch, for parallel and sequential builds using `make` and `ninja`, for all three platforms.

Results are represented in the following table:

| Platform | Build Type | Command | Elapsed Time (`master`) | Elapsed Time (`rebar-build`) |
|----------|------------|---------|-------------------------|------------------------------|
| MacOS    | clean | `make -j 8` | 2m51.213s | 0m33.763s |
| MacOS    | rebuild | `make -j 8` | 0m5.306s | 0m13.155s |
| MacOS    | clean | `ninja` | 2m41.689s | 0m30.022s |
| MacOS    | rebuild | `ninja` | 0m4.336s | 0m9.363s |
| MacOS    | clean | `make` | 5m48.742s | 1m27.505s |
| MacOS    | rebuild | `make` | 0m14.352s | 0m24.544s |
| | | | |
| FreeBSD    | clean | `make -j 8` | 1m5.755s | 0m34.545s |
| FreeBSD    | rebuild | `make -j 8` | 0m2.386s | 0m11.962s |
| FreeBSD    | clean | `ninja` | 0m57.580s | 0m29.580s |
| FreeBSD    | rebuild | `ninja` | 0m1.661s | 0m5.423s |
| FreeBSD    | clean | `make` | 3m32.387s | 1m11.465s |
| FreeBSD    | rebuild | `make` | 0m7.584s | 0m14.974s |
| | | | |
| Debian (virt) | clean | `make -j 8` | 1m39.644s | 0m34.479s |
| Debian (virt) | rebuild | `make -j 8` | 0m10.828s | 0m13.745s |
| Debian (virt) | clean | `ninja` | 2m55.802s | 0m38.964s |
| Debian (virt) | rebuild | `ninja` | 0m5.002s | 0m10.531s |
| Debian (virt) | clean | `make` | 9m54.355s | 1m51.219s |
| Debian (virt) | rebuild | `make` | 0m26.210s | 0m36.631s |

Summary:

* Clean builds are generally 2-3x faster using rebar than on `master`
* Rebuilds (with no changes) are roughly 2x slower using `rebar3`
* Ninja and parallel builds are roughly similar on clean builds, but Ninja outperforms on rebuilds.
* Clean sequential builds on `master` are extremely slow.  They are better with `rebar3`, as at least chunks of the Erlang code are built in parallel.  This is especially true of the erlang unit tests.

## Future Improvements

* The Erlang unit tests are now packaged into an AVM.  This may make them suitable for testing on ESP32, STM32, rp2040, etc.
* Using the `atomvm_rebar3_plugin` tees up support for the OTP Application behavior, as the plugin will generate `application.bin` files from the created `.app` files, which are used at runtime by the OTP Application controller to determine which applications to load.
* The Elixir builds should eventually use `mix` and the `ExAtomVM` plugin, but we have some work to do on the plugin to brin git up to parity.  We have left the Elixir builds as is.

## Notes

* MacOS
    * OS: 11.7.10
    * Hardware: Late 2014 iMac 4 GHz Quad-Core Intel Core i7 32G RAM
* FreeBSD
    * OS: `FreeBSD freenas.local 13.1-RELEASE-p7 FreeBSD 13.1-RELEASE-p7 n245428-4dfb91682c1 TRUENAS amd64`
    * Hardware: Intel(R) Xeon(R) CPU D-1518 @ 2.20GHz 8 core 32G RAM
* Debian
    * OS: `Linux debian 4.19.0-24-amd64 #1 SMP Debian 4.19.282-1 (2023-04-29) x86_64 GNU/Linux`
    * Virtualization: FreeBSD BHyve  4 cores assigned, 4GB RAM assigned
    * Host OS: See FreeBSD
    * Hardware: See FreeBSD

These changes are made under both the "Apache 2.0" and the "GNU Lesser General
Public License 2.1 or later" license terms (dual license).

SPDX-License-Identifier: Apache-2.0 OR LGPL-2.1-or-later