@UncleGrumpy 2023-12-06
I just did a test run with this PR and it looks like it is still building the C `PackBEAM` target. Should those build steps be removed?