@bettio 2024-01-15
> The alisp project has been folded into the eavmlib project, as we have no use case for building or deploying these modules separately.

Honestly my plan here is quite the opposite and I would further modularize libraries, such as I mention here #710.