@bettio 2024-01-15
Tested release workflows, and I've got the following error:
Status:
```
build-libraries
Process completed with exit code 2.
```

Log:
```
[ 27%] Linking C executable test-structs
[ 27%] Built target test-structs
[ 27%] Creating code_load packbeam ...
/bin/sh: 1: rebar3: not found
make[2]: *** [tests/code_load/CMakeFiles/code_load_packbeam.dir/build.make:71: tests/code_load/CMakeFiles/code_load_packbeam] Error 127
make[1]: *** [CMakeFiles/Makefile2:1095: tests/code_load/CMakeFiles/code_load_packbeam.dir/all] Error 2
make: *** [Makefile:136: all] Error 2
Error: Process completed with exit code 2.
```