@fadushin 2024-02-10
I am going to keep this branch open but close the issue, for a rethink of strategies.